function/requirements.txt
google-cloud-firestore==2.3.0
google-cloud-vision==2.4.2
google-cloud-translate==3.4.0
